update the `me` query's cache after `register`/`login` mutations
without this, once you register/login, the parts of the app that used
the `me` query would have the cached result which was previously `null`
and so it'd be out of sync.

We could invalidate the cache instead - it would make an extra request tho
and here we already have all the data we need.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>